Upravná aplikácia s oddelenou logikou. Vytvorené testy na spracovanie a kontrolu údajov.
