+++ Lekcia12_Cvicenie.sln
@@ -1,9 +1,11 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
-VisualStudioVersion = 17.12.35527.113 d17.12
+VisualStudioVersion = 17.12.35527.113
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lekcia12_Cvicenie", "Lekcia12_Cvicenie.csproj", "{CEFBEDD0-1D4C-4E98-9877-7BF76741B28D}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Lekcia13_Cvicenie", "..\Lekcia13_Cvicenie\Lekcia13_Cvicenie.csproj", "{EF412871-41ED-4F30-A43F-F98AB9F8BB4F}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -15,6 +17,10 @@ Global
 		{CEFBEDD0-1D4C-4E98-9877-7BF76741B28D}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CEFBEDD0-1D4C-4E98-9877-7BF76741B28D}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CEFBEDD0-1D4C-4E98-9877-7BF76741B28D}.Release|Any CPU.Build.0 = Release|Any CPU
+		{EF412871-41ED-4F30-A43F-F98AB9F8BB4F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{EF412871-41ED-4F30-A43F-F98AB9F8BB4F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{EF412871-41ED-4F30-A43F-F98AB9F8BB4F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{EF412871-41ED-4F30-A43F-F98AB9F8BB4F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
